requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
attrs==25.4.0
contourpy==1.3.3
coverage==7.13.4
cycler==0.12.1
execnet==2.1.2
fonttools==4.61.1
frozenlist==1.8.0
greenlet==3.3.2
idna==3.11
iniconfig==2.3.0
kiwisolver==1.4.9
matplotlib==3.10.8
multidict==6.7.1
numpy==2.4.2
packaging==26.0
pillow==12.1.1
pluggy==1.6.0
propcache==0.4.1
Pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
six==1.17.0
SQLAlchemy==2.0.48
typing_extensions==4.15.0
websockets==16.0
yarl==1.23.0
